requirements.txt
streamlit
Pillow
numpy
tensorflow
keras
scikit-learn
seaborn
pandas
matplotlib